主键与时间戳数据类型修改为bigint

+++ migrations/2020_01_16_171747_create_resource_table.php
@@ -12,7 +12,7 @@ class CreateResourceTable extends Migration
     public function up(): void
     {
         Schema::create('resource', function (Blueprint $table) {
-            $table->increments('id')
+            $table->bigIncrements('id')
                 ->comment('primary key');
             $table->string('key', 50)
                 ->unique()
@@ -45,11 +45,11 @@ class CreateResourceTable extends Migration
                 ->default(1)
                 ->unsigned()
                 ->comment('status');
-            $table->integer('create_time')
+            $table->bigIncrements('create_time')
                 ->default(0)
                 ->unsigned()
                 ->comment('create time');
-            $table->integer('update_time')
+            $table->bigIncrements('update_time')
                 ->default(0)
                 ->unsigned()
                 ->comment('update time');
